dockers/tesseract_docker/requirements.txt
flask
pytesseract
Pillow 
